Feature/productservice/facility setup

+++ productservice/src/main/java/com/naodab/productservice/controllers/CategoryController.java
@@ -5,12 +5,17 @@ import java.util.List;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.http.ResponseEntity;
 import jakarta.validation.Valid;
+
+import com.naodab.commonservice.constant.AppConstants;
+import com.naodab.commonservice.exception.AppException;
+import com.naodab.commonservice.exception.ErrorCode;
 import com.naodab.commonservice.response.ApiResponse;
 import com.naodab.productservice.dto.request.CategoryCreateRequest;
 import com.naodab.productservice.dto.response.CategoryResponse;
@@ -31,7 +36,13 @@ public class CategoryController {
 
   @PostMapping
   public ResponseEntity<ApiResponse<CategoryResponse>> createCategory(
+      @RequestHeader(value = AppConstants.JWT_CLAIM_ROLE, required = false) String role,
       @RequestBody @Valid CategoryCreateRequest request) {
+
+    if (role != null && !role.equals(AppConstants.ROLE_ADMIN)) {
+      throw new AppException(ErrorCode.FORBIDDEN);
+    }
+
     return ResponseEntity.ok(
         ApiResponse.<CategoryResponse>builder()
             .data(categoryService.createCategory(request))
@@ -72,7 +83,11 @@ public class CategoryController {
   }
 
   @DeleteMapping("/{id}")
-  public ResponseEntity<ApiResponse<Void>> deleteCategory(@PathVariable String id) {
+  public ResponseEntity<ApiResponse<Void>> deleteCategory(
+      @RequestHeader(value = AppConstants.JWT_CLAIM_ROLE, required = false) String role, @PathVariable String id) {
+    if (role != null && !role.equals(AppConstants.ROLE_ADMIN)) {
+      throw new AppException(ErrorCode.FORBIDDEN);
+    }
     categoryService.deleteCategory(id);
     return ResponseEntity.ok(
         ApiResponse.<Void>builder()
